+++ test/form/samples/render-chunk-plugin/_expected/system.js
@@ -1,0 +1,2 @@
+/* first plugin */
+/* second plugin */