+++ test/sourcemaps/samples/names-transformed-render-chunk/a.js
@@ -1,0 +1,4 @@
+var mangleMe = "this string should not be inlined".toLowerCase();
+export default function () {
+	assert.equal( mangleMe, 1 );
+}